add BatchNumberAPI, batch # & placebo to fda table
