+++ src/components/AssignBatchNumberButton.tsx
@@ -1,9 +1,9 @@
-import AssignBatchNumber from "@/api/assignBatchNumber";
+import { BatchNumberAPI } from "@/api/BatchNumberAPI";
 import { Button } from "@mui/material";
 
 export default function AssignBatchNumberButton()
 {
     return(
-        <Button variant='contained' onClick={() => {AssignBatchNumber()}}>Assign Batch Numbers</Button>
+        <Button variant='contained' onClick={() => {BatchNumberAPI.AssignBatchNumber()}}>Assign Batch Numbers</Button>
     )
 }
